+++ starkbank/__init__.py
@@ -1,4 +1,4 @@
-version = "2.34.0"
+version = "2.35.0"
 
 user = None
 language = "en-US"